1. Reorganize WLFU (use onNewQuery() to check rather than query id), fix small issues in SegmentCacheActor. 2. Connect query execution to front end.

+++ normal-cache/src/policy/BeladyCachingPolicy.cpp
@@ -247,7 +247,7 @@ std::string BeladyCachingPolicy::printLayoutAfterEveryQuery() {
 }
 
 void BeladyCachingPolicy::assertDecreasingOrderingOfSegmentKeys(const std::shared_ptr<std::vector<std::shared_ptr<SegmentKey>>>& segmentKeys) {
-  for (int i = 0; i < segmentKeys->size() - 1; i++) {
+  for (size_t i = 0; i < segmentKeys->size() - 1; i++) {
     auto key1 = segmentKeys->at(i);
     auto key2 = segmentKeys->at(i + 1);
     if (lessKeyValue(key1, key2)) {
